Edit functionality added and fixed delete bug
