+++ src/service/DataManager.js
@@ -3,7 +3,30 @@ class DataManager {
         return JSON.parse(localStorage.getItem('leetcode-memo'));
     }
 
-    saveData(data) {
+    saveData(isNew, item, data) {
+        if (isNew) {
+            data.push(item);
+        } else if (!isNew) {
+            data = data.map(t => {
+                let temp = Object.assign({}, t);
+                if (temp.id === item.id) {
+                    temp.title = item.title;
+                    temp.link = item.link;
+                    temp.difficulty = item.difficulty;
+                    temp.timeSpent = item.timeSpent;
+                    temp.category = item.category;
+                    temp.note = item.note;
+                    temp.personalDifficulty = item.personalDifficulty;
+                    temp.isFavourite = item.isFavourite;
+                }
+                return temp;
+            })
+        }
+
+        return this.setLocalStorage(data);
+    }
+
+    setLocalStorage(data) {
         return localStorage.setItem("leetcode-memo", JSON.stringify(data));
     }
 
@@ -27,7 +50,7 @@ class DataManager {
         let data = this.getData();
         if (data.length == 1) return this.deleteData();
         
-        return this.saveData(data.filter(item => item.id != id));
+        return this.setLocalStorage(data.filter(item => item.id != id));
     }
 }
